Currently selected language now updated in template list.
List of languages in drop down menu is now taken from the template.
The dialog maintains the current selected language. When a different
language is selected from the drop down menu the selected language is
changed for all templates. Need to handle the case where the selected
language is not supported by the selected template.